requirements.txt
openai>=1.99.0
pandas>=2.2.0
openpyxl>=3.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
tenacity>=9.0.0
tqdm>=4.66.0
networkx>=3.3
rank-bm25>=0.2.2
scikit-learn>=1.5.0
numpy>=1.26.0
orjson>=3.10.0
